pyproject.toml
[project]
name = "ai-hindi-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.5",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "sqlalchemy>=2.0.36",
    "supabase>=2.10.0",
    "uvicorn>=0.32.1",
    "websockets>=13.1",
]
